+++ src/app.ts
@@ -59,5 +59,8 @@ export default (tsconfigPath: string, files?: string[]): Analysis => {
   const args = extractOptionsFromFiles(files);
   const tsConfig = loadTsConfig(tsconfigPath, args.tsFiles);
 
-  return analyze(parseFiles(dirname(tsconfigPath), tsConfig), args.options);
+  return analyze(
+    parseFiles(dirname(tsconfigPath), tsConfig, args.options),
+    args.options,
+  );
 };