+++ src/types.ts
@@ -39,6 +39,7 @@ export interface TsConfig {
 
 export interface ExtraCommandLineOptions {
   exitWithCount?: boolean;
+  includeDeclarationFiles?: boolean;
   pathsToIgnore?: string[];
   showLineNumber?: boolean;
 }